#182 Setup the arm, and Go indicator
Set it all up with a boolean method which may be bad, may want to change the label in the controller rather than the information view? Also may want something other than a label. Another thing to consider is after pressing arm, a simulation does not run? Why was this?